Add proc_spawn3 and proc_exec4, avoiding the \n delimited argv and envp from older syscalls
