+++ lib/wasix/src/syscalls/wasix/mod.rs
@@ -38,6 +38,7 @@ mod port_unbridge;
 mod proc_exec;
 mod proc_exec2;
 mod proc_exec3;
+mod proc_exec4;
 mod proc_exit2;
 mod proc_fork;
 mod proc_fork_env;
@@ -50,6 +51,7 @@ mod proc_signals_sizes_get;
 mod proc_snapshot;
 mod proc_spawn;
 mod proc_spawn2;
+mod proc_spawn3;
 mod reflect_signature;
 mod resolve;
 mod sched_yield;
@@ -130,6 +132,7 @@ pub use port_unbridge::*;
 pub use proc_exec::*;
 pub use proc_exec2::*;
 pub use proc_exec3::*;
+pub use proc_exec4::*;
 pub use proc_exit2::*;
 pub use proc_fork::*;
 pub use proc_fork_env::*;
@@ -142,6 +145,7 @@ pub use proc_signals_sizes_get::*;
 pub use proc_snapshot::*;
 pub use proc_spawn::*;
 pub use proc_spawn2::*;
+pub use proc_spawn3::*;
 pub use reflect_signature::*;
 pub use resolve::*;
 pub use sched_yield::*;
