+++ lib/wasix/tests/wasm_tests/proc_exec4_newline_arg/proc_exec4_newline_arg_child.c
@@ -1,0 +1,8 @@
+#include <assert.h>
+#include <string.h>
+
+int main(int argc, char** argv) {
+  assert(argc >= 2);
+  assert(strcmp(argv[1], "line1\nline2") == 0);
+  return 0;
+}
