+++ lib/wasix/tests/wasm_tests/proc_exec4_newline_env/build.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+$CC proc_exec4_newline_env_child.c -o proc_exec4_newline_env_child.wasm
+$CC main.c -o main
