+++ lib/wasix/tests/wasm_tests/proc_exec4_newline_env/proc_exec4_newline_env_child.c
@@ -1,0 +1,13 @@
+#include <assert.h>
+#include <stdlib.h>
+#include <string.h>
+
+int main(int argc, char** argv) {
+  (void)argc;
+  (void)argv;
+
+  const char* v = getenv("NEWLINE_ENV");
+  assert(v != NULL);
+  assert(strcmp(v, "value\nwith\nnewlines") == 0);
+  return 0;
+}
