+++ lib/wasix/tests/wasm_tests/proc_spawn3_newline_arg/build.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+$CC main.c -o main
